+++ frontend/packages/web-app/src/views/Highlight/Index.vue
@@ -1,81 +1,22 @@
 <!-- @format -->
 
 <script lang="ts" setup>
-import { ref, computed, onMounted, onUnmounted } from 'vue'
-import { windowManager } from '@/platform'
-import { PickShortCuts, PickMode, TipPosition } from './config'
+import { ref } from 'vue'
 import ConfigProvider from '@/components/ConfigProvider/index.vue'
-import { RpaHighlight } from '@/api/highlight'
+import { useHighlight } from './hooks/useHighlight'
 
-const ws = ref<WebSocket | null>(null)
-const highlightRect = ref({ x: 0, y: 0, width: 0, height: 0 })
 const tooltipRef = ref<HTMLElement | null>(null)
-const mousePos = ref({ x: 0, y: 0 })
-const tooltipVisible = ref(true)
-const pickMode = ref('')
-const appName = ref('')
-const tagName = ref('')
-const dpr = window.devicePixelRatio || 1
-
-const tooltipPos = computed(() => {
-  const margin = 300
-  const mouse = mousePos.value
-  if (mouse.x < margin && mouse.y < margin) {
-    return 'rightBottom'
-  }
-  if (mouse.x > (screen.width  - margin) * dpr && mouse.y > (screen.height - margin) * dpr) {
-    return  'leftTop'
-  }
-  return 'rightBottom'
-})
-
-const tagPosition = computed(() => {
-  const rect = highlightRect.value
-  const margin = 60
-  if (rect.y < margin) {
-    return 'bottom'
-  }
-    else {
-      return 'top'
-    }
-})
-
-pickMode.value = PickMode.NORMAL
-
-const shortcuts = computed(() => PickShortCuts[pickMode.value] || [])
-onMounted(() => {
-  RpaHighlight.create(() => {
-    
-    RpaHighlight.bindMessage((data) => {
-      const op = data.Operation
-      if (op === 'start') {
-        windowManager.showWindow()
-        if (data.Type) pickMode.value = data.Type
-      } else if (op === 'hide') {
-        windowManager.hideWindow()
-        highlightRect.value = { x: 0, y: 0, width: 0, height: 0 }
-        tagName.value = ''
-      } else if (op === 'draw') {
-        const boxes = data.Boxes
-        if (boxes && boxes.length > 0) {
-          const box = boxes[0]
-          highlightRect.value = {
-            x: box.Left,
-            y: box.Top,
-            width: (box.Right - box.Left) / dpr,
-            height: (box.Bottom - box.Top) / dpr,
-          }
-          if (box.Msg !== undefined) tagName.value = box.Msg
-        }
-      }
-      windowManager.setWindowAlwaysOnTop(true)
-    })
-  })
-})
-
-onUnmounted(() => {
-  ws.value?.close()
-})
+const {
+  dpr,
+  highlightRect,
+  mousePos,
+  appName,
+  tagName,
+  tooltipVisible,
+  tooltipPos,
+  tagPosition,
+  shortcuts,
+} = useHighlight()
 </script>
 
 <template>
